requirements.txt
Flask==2.2.5
Flask_Cors==4.0.0
Pillow==10.1.0
